// src/enum/project.enum.tsx
export enum EFilterStatusNumber {
  ACTIVE = 0,
  DEACTIVE = 1,
  ALL = 2
}

export enum ETypeMember {
  ALL = 'ALL',
  STAFF = 'STAFF',
  INTERNSHIP = 'INTERNSHIP',
  COLLABORATOR = 'COLLABORATOR'
}

export enum EProjectType {
  TM = 'T&M',
  FIXEDPRICE = 'Fixed Price',
  NONBILLABLE = 'Non-Billable',
  ODC = 'ODC',
  PRODUCT = 'Product',
  TRAINING = 'Training',
  NOSALLARY = 'NoSalary'
}

export enum EProjectTypeShortName {
  TM = 'T&M',
  FIXEDPRICE = 'FF',
  NONBILLABLE = 'NB',
  ODC = 'ODC',
  PRODUCT = 'PD',
  TRAINING = 'TN',
  NOSALLARY = 'NS'
}

export enum EProjectTypeValue {
  TM = 0,
  FIXEDPRICE = 1,
  NONBILLABLE = 2,
  ODC = 3,
  PRODUCT = 4,
  TRAINING = 5,
  NOSALLARY = 6
}

export enum EViewTimeSelect {
  WEEK = 'Week',
  MONTH = 'Month',
  QUATER = 'Quater',
  YEAR = 'Year',
  ALL_TIME = 'All Time',
  CUSTOM_TIME = 'Custom Time'
}